Fixes a bug in find_best_cycle() Introduces a new shift-swap based strategy. Performance is greatly improved.
This commit fixes a bug which prevented the algorithm
capitalise on the discovery of a new more efficient
path. It also strategically introduces the idea of a
portfolio (set of tuples)
of cycles taken into account so far.
The algorithm now calls shift_cities if after
N swaps an old configuration comes back.